Restored the work on ConsoleApp
-This file is remaining in src until we find a way to refactor it that
works for everyone to still access it
-Currently, mainloop is called from SSHTest. Just comment out that line
if you don't want to run the console app.
-PLEASE don't merge with master without testing first
